// src/client/types.gen.ts
export type ColorEnum = 'navy' | 'blue' | 'aqua' | 'teal' | 'olive' | 'green' | 'lime' | 'yellow' | 'orange' | 'red' | 'maroon' | 'fuchsia' | 'purple' | 'black' | 'gray' | 'silver';
// ...
export type StateEnum = 0 | 10 | 20 | 30 | 100;
// ...
export type Unit = {
    readonly translation: string;
    readonly language_code: string;
    source: Array<string>;
    previous_source?: string;
    target: Array<string>;
    id_hash: number;
    readonly content_hash: number;
    location?: string;
    context?: string;
    note?: string;
    flags?: string;
    labels: Array<UnitLabels>;
    state?: StateEnum;
    readonly fuzzy: boolean;
    readonly translated: boolean;
    readonly approved: boolean;
    position: number;
    readonly has_suggestion: boolean;
    readonly has_comment: boolean;
    readonly has_failing_check: boolean;
    num_words?: number;
    readonly source_unit: string;
    priority?: number;
    readonly id: number;
    readonly web_url: string;
    readonly url: string;
    /**
     * Additional explanation to clarify meaning or usage of the string.
     */
    explanation?: string;
    /**
     * Translation flags
     * Additional comma-separated flags to influence Weblate behavior.
     */
    extra_flags?: string;
    readonly pending: boolean;
    readonly timestamp: string;
    readonly last_updated: string;
};

export type UnitFlatLabels = {
    readonly id: number;
    /**
     * Label name
     */
    name: string;
    /**
     * Label description
     */
    description?: string;
    color?: ColorEnum;
};

export type UnitLabels = {
    readonly id: number;
    /**
     * Label name
     */
    name: string;
    /**
     * Label description
     */
    description?: string;
    color?: ColorEnum;
};